requirements.txt
# Python 3.12+
# Core
numpy>=1.26.0
pandas>=2.0.0
pyarrow>=14.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
joblib>=1.3.0
requests>=2.28.0

# Implicit ALS
implicit>=0.7.0

# Price estimation (gradient boosting)
lightgbm>=4.0.0

# Config & experiment tracking
pyyaml>=6.0
mlflow>=2.9.0

# Optional: W&B alternative
# wandb>=0.15.0
